refactor(desktop): 主窗口生命周期抽到 window.ts
把建窗 + 窗口状态（尺寸/最大化）载入/回写从 index.ts 抽到 createWindowManager（window.ts）：载入窗口状态、
create(splash)（恢复尺寸 + 自绘标题栏 + 外链路由 + 首帧显示/关 splash）、resize/move/close 防抖回写。
windowState 内聚到管理器、不再是 index 模块级可变全局。index 改用 windowManager.create()；顺带移除随之
不再使用的 shell / __dirname / fileURLToPath。纯搬运、行为不变。
